#602: removed v2 of download and opds and updated exsisting services
