+++ src/main/scala/ScalatraBootstrap.scala
@@ -18,9 +18,7 @@ class ScalatraBootstrap extends LifeCycle {
     context.mount(ComponentRegistry.searchController, BookApiProperties.SearchPath, "search")
     context.mount(ComponentRegistry.searchControllerV2, BookApiProperties.SearchPathV2, "searchV2")
     context.mount(ComponentRegistry.opdsController, BookApiProperties.OpdsPath, "opds")
-    context.mount(ComponentRegistry.opdsControllerV2, BookApiProperties.OpdsPath, "opdsV2")
     context.mount(ComponentRegistry.downloadController, BookApiProperties.DownloadPath, "downloads")
-    context.mount(ComponentRegistry.downloadControllerV2, BookApiProperties.DownloadPathV2, "downloadsV2")
     context.mount(ComponentRegistry.languageController, BookApiProperties.LanguagesPath, "languages")
     context.mount(ComponentRegistry.languageControllerV2, BookApiProperties.LanguagesPathV2, "languagesV2")
     context.mount(ComponentRegistry.levelController, BookApiProperties.LevelsPath, "levels")
